comments+orders+pagination

+++ BackendApp/reposio/cart_itemsDal.js
@@ -40,13 +40,12 @@ class CartModel {
 
     async getAll(user_id) {
         try {
-            const cart_items = await pool('cart_items').where({ user_id: user_id }).select();
-            const prodIds = cart_items.map(item => item.prod_id);
+            const cart = await pool('production as p')
+                .select('p.*', 'ci.quantity')
+                .join('cart_items as ci', 'p.id', 'ci.prod_id')
+                .where('ci.user_id', user_id);
 
-            const cart = await pool('production').whereIn( 'id', prodIds ).select();
-            const result = [cart_items, cart]; 
-            
-            return result;
+            return cart;
         } catch (err) {
             console.error('Error fetching cart_items', err);
             throw err;
@@ -58,9 +57,7 @@ class CartModel {
     async delete(user_id, prod_id) {
         try {
             const query = pool('cart_items');
-            await query
-                .where({ user_id: user_id, prod_id: prod_id })
-                .delete();
+            await query.where({ user_id: user_id, prod_id: prod_id }).delete();
         } catch (err) {
             console.error('Error fetching cart_item by ID', err);
             throw err;
